+++ src/packages/array_basic.rs
@@ -3,7 +3,7 @@
 use crate::any::{Dynamic, Variant};
 use crate::def_package;
 use crate::engine::{Array, Engine};
-use crate::module::FuncReturn;
+use crate::module::{FuncReturn, Module};
 use crate::parser::{ImmutableString, INT};
 use crate::result::EvalAltResult;
 use crate::token::Position;
@@ -25,7 +25,11 @@ fn ins<T: Variant + Clone>(list: &mut Array, position: INT, item: T) -> FuncRetu
     }
     Ok(())
 }
-fn pad<T: Variant + Clone>(engine: &Engine, args: &mut [&mut Dynamic]) -> FuncReturn<()> {
+fn pad<T: Variant + Clone>(
+    engine: &Engine,
+    _: &Module,
+    args: &mut [&mut Dynamic],
+) -> FuncReturn<()> {
     let len = *args[1].downcast_ref::<INT>().unwrap();
 
     // Check if array will be over max size limit