+++ src/settings.rs
@@ -218,8 +218,8 @@ impl Engine {
     /// let mut engine = Engine::new();
     ///
     /// // Register a custom operator called 'foo' and give it
-    /// // a precedence of 140 (i.e. between +|- and *|/).
-    /// engine.register_custom_operator("foo", 140).unwrap();
+    /// // a precedence of 160 (i.e. between +|- and *|/).
+    /// engine.register_custom_operator("foo", 160).unwrap();
     ///
     /// // Register a binary function named 'foo'
     /// engine.register_fn("foo", |x: i64, y: i64| (x * y) - (x + y));